+++ src/launcher/ipc_server/mod.rs
@@ -317,131 +317,4 @@ mod test {
         // design strives for.
         eval_result!(event_sender.send(::launcher::ipc_server::events::ExternalEvent::Terminate));
     }
-
-    #[derive(Debug)]
-    struct HandshakeRequest {
-        pub end_point: String,
-        pub data: HandshakePayload,
-    }
-
-    #[derive(Debug)]
-    struct HandshakePayload {
-        pub launcher_string: String,
-        pub nonce: [u8; 24],
-        pub public_encryption_key: [u8; 32],
-    }
-
-    impl ::rustc_serialize::json::ToJson for HandshakePayload {
-        fn to_json(&self) -> ::rustc_serialize::json::Json {
-            use ::rustc_serialize::base64::ToBase64;
-
-            let mut tree = ::std::collections::BTreeMap::new();
-            let config = ::config::get_base64_config();
-            let base64_nonce = (&self.nonce).to_base64(config);
-            let base64_pub_encryption_key = (&self.public_encryption_key).to_base64(config);
-
-            if tree.insert("launcher_string".to_string(), self.launcher_string.to_json()).is_some() {
-                error!("Json Conversion error -- HandshakePayload -- launcher_string");
-            }
-            if tree.insert("nonce".to_string(), base64_nonce.to_json()).is_some() {
-                error!("Json Conversion error -- HandshakePayload -- nonce");
-            }
-            if tree.insert("public_encryption_key".to_string(), base64_pub_encryption_key.to_json()).is_some() {
-                error!("Json Conversion error -- HandshakePayload -- public_encryption_key");
-            }
-
-            ::rustc_serialize::json::Json::Object(tree)
-        }
-    }
-
-
-    impl ::rustc_serialize::json::ToJson for HandshakeRequest {
-        fn to_json(&self) -> ::rustc_serialize::json::Json {
-            use ::rustc_serialize::base64::ToBase64;
-
-            let mut tree = ::std::collections::BTreeMap::new();
-            let config = ::config::get_base64_config();
-
-            if tree.insert("end_point".to_string(), self.end_point.to_json()).is_some() {
-                error!("Json Conversion error -- HandshakeRequest -- endpoint");
-            }
-            if tree.insert("data".to_string(), self.data.to_json()).is_some() {
-                error!("Json Conversion error -- HandshakeRequest -- data");
-            }
-
-            ::rustc_serialize::json::Json::Object(tree)
-        }
-    }
-
-    #[test]
-    fn application_handshake() {
-        use ::rustc_serialize::json::ToJson;
-
-        let client = ::std
-                     ::sync
-                     ::Arc::new(::std
-                                ::sync
-                                ::Mutex::new(eval_result!(::safe_core::utility::test_utils::get_client())));
-
-        let (_raii_joiner_0, event_sender) = eval_result!(IpcServer::new(client));
-
-        let (tx, rx) = ::std::sync::mpsc::channel();
-        eval_result!(event_sender.send(::launcher::ipc_server::events::ExternalEvent::GetListenerEndpoint(tx)));
-        let listener_ep = eval_result!(rx.recv());
-
-        let mut stream = eval_result!(::std::net::TcpStream::connect(&listener_ep[..]));
-
-        let app_id = ::routing::NameType(eval_result!(::safe_core::utility::generate_random_array_u8_64()));
-        let dir_id = ::routing::NameType(eval_result!(::safe_core::utility::generate_random_array_u8_64()));
-        let directory_key = ::safe_nfs::metadata::directory_key::DirectoryKey::new(dir_id,
-                                                                                   10u64,
-                                                                                   false,
-                                                                                   ::safe_nfs::AccessLevel::Private);
-        let activation_details = ::launcher::ipc_server::events::event_data::ActivationDetail {
-            nonce            : "mock_nonce_string".to_string(),
-            app_id           : app_id,
-            app_root_dir_key : directory_key,
-            safe_drive_access: false,
-        };
-        let activate_event = ::launcher::ipc_server::events::ExternalEvent::AppActivated(Box::new(activation_details));
-        event_sender.send(activate_event);
-
-        let _raii_joiner_1 = ::safe_core
-                             ::utility
-                             ::RAIIThreadJoiner
-                             ::new(eval_result!(::std
-                                                ::thread
-                                                ::Builder::new().name("TCPClientThread".to_string())
-                                                                .spawn(move || {
-                use rustc_serialize::base64::ToBase64;
-
-                let mut ipc_stream = eval_result!(::launcher
-                                                  ::ipc_server
-                                                  ::ipc_session
-                                                  ::stream
-                                                  ::IpcStream::new(eval_result!(stream.try_clone())));
-                let app_nonce = ::sodiumoxide::crypto::box_::gen_nonce();
-                let (app_public_key, app_seccret_key) = ::sodiumoxide::crypto::box_::gen_keypair();
-                let payload = HandshakePayload {
-                    launcher_string      : "mock_nonce_string".to_string(),
-                    nonce                : app_nonce.0,
-                    public_encryption_key: app_public_key.0,
-                };
-                let request = HandshakeRequest {
-                    end_point: "safe-api/v1.0/handshake/authenticate-app".to_string(),
-                    data     : payload,
-                };
-
-                let json_obj = request.to_json();
-                let serialised_data = eval_result!(::safe_core::utility::serialise(&json_obj.to_string()));
-
-                ipc_stream.write(serialised_data);
-
-                ::std::thread::sleep_ms(1000);
-                let mut buffer = [0; 1024];
-                assert_eq!(eval_result!(stream.read(&mut buffer)), 1024);
-        })));
-        ::std::thread::sleep_ms(3000);
-        eval_result!(event_sender.send(::launcher::ipc_server::events::ExternalEvent::Terminate));
-    }
 }
